sglang dcu 0.5.10rc0 base bash(48739c3a)

+++ python/sglang/srt/layers/attention/base_attn_backend.py
@@ -90,7 +90,7 @@ class AttentionBackend(ABC):
     ):
         """Run forward on an attention layer."""
         if forward_batch.forward_mode.is_idle():
-            return q.new_empty(q.shape[0], layer.tp_q_head_num * layer.v_head_dim)
+            return q.new_empty(q.shape[0], layer.tp_q_head_num * layer.v_head_dim,dtype=torch.bfloat16)
         elif forward_batch.forward_mode.is_decode():
             return self.forward_decode(
                 q,
